backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
langgraph
openai
pydantic